requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite
langchain-mcp-adapters
aiosqlite
faiss-cpu
pypdf
docx2txt
unstructured
python-pptx
httpx
mcp[cli]
